5758 - Organize validations/tables folder
