+++ src/server/controller/cycleData/validations/tables/shouldSkipValidationFormula.ts
