+++ src/tools/validations/validateAllRunner.ts
@@ -6,7 +6,7 @@ import { Promises } from 'utils/promises'
 import { AreaController } from 'server/controller/area'
 import { AssessmentController } from 'server/controller/assessment'
 import { validateCountryDescriptions } from 'server/controller/cycleData/validations/descriptions/validateCountryDescriptions'
-import { validateCountryTables } from 'server/controller/cycleData/validations/tables/validateCountryTables'
+import { validateCountryTables } from 'server/controller/cycleData/validations/tables/validateCountryTables/validateCountryTables'
 import { BaseProtocol, DB } from 'server/db/db'
 import { DescriptionRepository } from 'server/db/repository/assessmentCycle/descriptions'
 import { Logger } from 'server/utils/logger'
